Display context about the line of test failures.
For example, when displaying the error:

```
tagged_template.js: Error: Unexpected token "string-2" for "`" at 58:11. Expected "string"
```

It will follow it up with the line where the error occured, and a carrot indicating the place in the line:

```
    html`\${10}`;
               ^
```

Also import the html language, as the JS and TS tests now require it.