Adds basic Doxygen grammar for C/C++ comments.
Adds a Doxygen grammar that extends the existing Javadoc grammar. The new grammar redefines the defTag to support the optional parameter direction ("in", "out", or "in,out") and to highlight the formula tags "@f$", "@f[", and "@f]".

Updates the C (and C++) grammar to use "comment" from doxygen.grammar (instead of javadoc.grammar).