[py] Allow \n after in in *-comprehensions
This is not the most elegant fix, but it seems to work.

I would love to understand why the MaybePattern(in) lookahead fails in
the presence of a EOL. I suspect that somehow maybeSkipNewLine works
differently during lookahead (is the context right?), but I'm not sure
how to go about verifying that.